Improved: Code by relocating Party-related functionality from the util module to a dedicated module in store and updating PartyService accordingly.(#121)
