+++ src/services/UtilService.ts
@@ -343,14 +343,6 @@ const fetchCarrierPartyIds = async (query: any): Promise <any>  => {
   });
 }
 
-const fetchPartyInformation = async (query: any): Promise <any>  => {
-  return api({
-    url: "performFind",
-    method: "get",
-    params: query
-  });
-}
-
 const fetchShipmentMethodTypeDesc = async (query: any): Promise <any>  => {
   return api({
     url: "performFind",
@@ -693,7 +685,6 @@ export const UtilService = {
   fetchGiftCardItemPriceInfo,
   fetchOrderAdjustments,
   fetchOrderShipGroupInfo,
-  fetchPartyInformation,
   fetchPicklistInformation,
   fetchProductStores,
   fetchRejectReasonEnumTypes,
