+++ src/store/RootState.ts
@@ -1,5 +1,6 @@
 import OrderState from "./modules/order/OrderState";
 import UtilState from "./modules/util/UtilState";
+import PartyState from "./modules/party/PartyState";
 import TransferOrderState from "./modules/transferorder/TransferOrderState";
 import CarrierState from "./modules/carrier/CarrierState"
 import RejectionState from "./modules/rejection/RejectionState"
@@ -9,6 +10,7 @@ export default interface RootState {
     user: any;
     product: any;
     util: UtilState;
+    party: PartyState;
     order: OrderState;
     transferOrder: TransferOrderState,
     carrier: CarrierState,
