+++ src/store/modules/party/PartyState.ts
@@ -1,0 +1,3 @@
+export default interface PartyState {
+    partyNames: any
+}
