+++ src/store/modules/party/actions.ts
@@ -1,0 +1,61 @@
+import { ActionTree } from 'vuex'
+import RootState from '@/store/RootState'
+import PartyState from './PartyState'
+import * as types from './mutation-types'
+import { PartyService } from '@/services/PartyService'
+import logger from '@/logger'
+import { hasError } from '@/adapter'
+
+const actions: ActionTree<PartyState, RootState> = {
+    async fetchPartyInformation({ commit, state }, partyIds) {
+    let partyInformation = JSON.parse(JSON.stringify(state.partyNames))
+    const cachedPartyIds = Object.keys(partyInformation);
+    const ids = partyIds.filter((partyId: string) => !cachedPartyIds.includes(partyId))
+
+    if(!ids.length) return partyInformation;
+
+    try {
+      const payload = {
+        "inputFields": {
+          "partyId": ids,
+          "partyId_op": "in"
+        },
+        "fieldList": ["firstName", "middleName", "lastName", "groupName", "partyId"],
+        "entityName": "PartyNameView",
+        "viewSize": ids.length
+      }
+
+      const resp = await PartyService.fetchPartyInformation(payload);
+
+      if(!hasError(resp)) {
+        const partyResp = {} as any
+        resp.data.docs.map((partyInformation: any) => {
+
+          let partyName = ''
+          if(partyInformation.groupName) {
+            partyName = partyInformation.groupName
+          } else {
+            partyName = [partyInformation.firstName, partyInformation.lastName].join(' ')
+          }
+
+          partyResp[partyInformation.partyId] = partyName
+        })
+
+        partyInformation = {
+          ...partyInformation,
+          ...partyResp
+        }
+
+        commit(types.UTIL_PARTY_NAMES_UPDATED, partyInformation)
+      } else {
+        throw resp.data
+      }
+    } catch(err) {
+      logger.error('Error fetching party information', err)
+    }
+
+    return partyInformation;
+  }
+}
+
+export default actions;
