+++ src/store/modules/party/getters.ts
@@ -1,0 +1,11 @@
+import { GetterTree } from "vuex";
+import PartyState from './PartyState'
+import RootState from "@/store/RootState";
+
+const getters: GetterTree<PartyState, RootState> = {
+    getPartyName: (state) => (partyId: string) => {
+        return state.partyNames[partyId] ? state.partyNames[partyId] : ''
+      }
+}
+
+export default getters;
