+++ src/store/modules/party/mutation-types.ts
@@ -1,0 +1,2 @@
+import { SN_UTIL } from "../util/mutation-types"
+export const UTIL_PARTY_NAMES_UPDATED = SN_UTIL + '/PARTY_NAMES_UPDATED'
