+++ src/store/modules/transferorder/actions.ts
@@ -209,7 +209,7 @@ const actions: ActionTree<TransferOrderState, RootState> = {
         while(productIds.length) {
           productIdBatches.push(productIds.splice(0, batchSize))
         }
-        await Promise.all([productIdBatches.map((productIds) => this.dispatch('product/fetchProducts', { productIds })), this.dispatch('util/fetchPartyInformation', [shipmentCarriers?.[0].carrierPartyId,]), this.dispatch('util/fetchShipmentMethodTypeDesc', [shipmentCarriers?.[0].shipmentMethodTypeId])])
+        await Promise.all([productIdBatches.map((productIds) => this.dispatch('product/fetchProducts', { productIds })), this.dispatch('party/fetchPartyInformation', [shipmentCarriers?.[0].carrierPartyId,]), this.dispatch('util/fetchShipmentMethodTypeDesc', [shipmentCarriers?.[0].shipmentMethodTypeId])])
       }
 
     } catch (err: any) {
