+++ src/store/modules/util/getters.ts
@@ -6,9 +6,6 @@ const getters: GetterTree <UtilState, RootState> = {
   getRejectReasons(state) {
     return state.rejectReasons ? state.rejectReasons : []
   },
-  getPartyName: (state) => (partyId: string) => {
-    return state.partyNames[partyId] ? state.partyNames[partyId] : ''
-  },
   getShipmentMethodDesc: (state) => (shipmentMethodId: string) => {
     return state.shipmentMethodTypeDesc[shipmentMethodId] ? state.shipmentMethodTypeDesc[shipmentMethodId] : shipmentMethodId
   },
