+++ src/store/modules/util/mutation-types.ts
@@ -1,6 +1,5 @@
 export const SN_UTIL = 'util'
 export const UTIL_REJECT_REASONS_UPDATED = SN_UTIL + '/REJECT_REASONS_UPDATED'
-export const UTIL_PARTY_NAMES_UPDATED = SN_UTIL + '/PARTY_NAMES_UPDATED'
 export const UTIL_SHIPMENT_METHODS_UPDATED = SN_UTIL + '/SHIPMENT_METHODS_UPDATED'
 export const UTIL_SHIPMENT_BOXES_UPDATED = SN_UTIL + '/SHIPMENT_BOXES_UPDATED'
 export const UTIL_FACILITY_TYPE_UPDATED = SN_UTIL + '/FACILITY_TYPE_UPDATED'
