+++ src/views/TransferShipmentReview.vue
@@ -169,7 +169,7 @@
         productIdentificationPref: 'user/getProductIdentificationPref',
         productStoreShipmentMethCount: 'util/getProductStoreShipmentMethCount',
         getShipmentMethodDesc: 'util/getShipmentMethodDesc',
-        getPartyName: 'util/getPartyName',
+        getPartyName: 'party/getPartyName',
       }),
     },
     methods: {
